rename files appropriately
